+++ src/convertFile/nodes.ts
@@ -1,0 +1,29 @@
+export const space = {
+  type: 'space',
+  value: ' ',
+};
+
+export const newLine = {
+  type: 'space',
+  value: '\n',
+};
+
+export const semicolon = {
+  type: 'punctuation',
+  value: ';',
+};
+
+export const comma = {
+  type: 'punctuation',
+  value: ',',
+};
+
+export const atValue = {
+  type: 'atkeyword',
+  value: 'value',
+};
+
+export const fromNode = {
+  type: 'identifier',
+  value: 'from',
+};