+++ src/convertFile/updateAtRules/index.ts
@@ -1,0 +1,14 @@
+import recurse from '../recurse';
+import updateRule from './updateRule';
+
+const updateAtRules = (data: Content.DataNode): Content.DataNode => {
+  const { node: { type }} = data;
+  if (type === 'atrule') {
+    updateRule(data);
+  }
+
+  recurse(data, 'updateAtRules');
+  return data;
+};
+
+export default updateAtRules;