+++ src/types/content.d.ts
@@ -1,8 +1,22 @@
-import { Context } from "mocha";
+import { Node } from 'scss-parser';
 
-export interface ParsedLine {
-  text: string;
-  isValue: boolean;
+export interface Rule {
+  astLocation: number,
+  node: Node
+}
+
+export interface NewVariable {
+  dictionaryEntry: string;
+  variable?: string;
+}
+
+export interface NewVariables {
+  [library: string]: string[];
+}
+
+export interface DataNode {
+  newVariables: NewVariables;
+  node: Node;
 }
 
 export as namespace Content;