+++ src/utils/index.ts
@@ -1,7 +1,4 @@
-export { default as ensureEmptyLine } from './ensureEmptyLine';
-export { default as match } from './match';
+export { default as findIndexFrom } from './findIndexFrom';
 export { default as moveCursor } from './moveCursor';
-export { default as organize } from './organize';
-export { default as removeLeadingBlanks } from './removeLeadingBlanks'
 export { default as startup } from './startup';
-export { default as updateText } from './updateText';
+export { default as updateText } from '../convertFile/updateText';