+++ docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the `tkww-assistant` extension will be documented in this file.
 
+## [0.1.0] - 2020-11-01
+### Changed
+- Change codebase to use abstract syntax tree for more stability
+- Changed `Convert to TKWW Union CSS Variables`  command to `TKWW: Convert File to Union Variables`.
+
 ## [0.0.1] - 2020-10-31
 ### Added
 - Initial release.